fix some ddl

+++ selfouettie-server/src/main/resources/db/migration/V2.3.0__create_issue_events_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE issue_events (
   event_id BIGINT UNSIGNED PRIMARY KEY,
   user_name VARCHAR(255) NOT NULL,
   repository_id BIGINT UNSIGNED NOT NULL,
-  issue_number BIGINT UNSIGNED NOT NULL,
+  issue_number INT UNSIGNED NOT NULL,
   action ENUM ('opened','edited','deleted','transferred','closed','reopened','assigned','unassigned','labeled','unlabeled','milestoned','demilestoned'),
   created_at BIGINT UNSIGNED,
   FOREIGN KEY(event_id) REFERENCES events(id),
